fix: brute-force approach.
The reason for this was that we do not have a class structure for the
form data: it is essentially an arbitrary json tree. Creating a custom
de/serializer seems tricky: you can't extend `JsonSerializer<String>`,
only `<JsonNode>` which I couldn't get to gracefully apply the default
de/serialization to non-text nodes.

https://github.com/FasterXML/jackson-databind/issues/3347
"when reading a tree (JsonNode, or more generically, TreeNode), other
types are not considered and this is by design. Any String-valued
properties are read as TextNodes, not as Strings.
Put another way, tree nodes (JsonNode) are not POJOs and do not have
typed properties in the same sense.
This means that you cannot (re)configure handling the same way you would
for POJO values.

You can override deserializer for JsonNode itself and customize
deserialization; however, there are no convenient ways to customize
deserialization of JsonNode scalar values. So it is also not possible
currently to override deserializer for TextNode."

TIS21-4623