fix: container memory bump
Some paketo-buildpacks versions have changed, which might be triggering
errors in the deployment to staging. This is a speculative patch pending
a proper investigation into how we can control those versions (and
indeed whether they are the root cause for the memory exception logs
noted in the failed deployments).

TIS21-4667