refactor: upgrade to Spring Boot 3.x.x
Upgrade to Spring Boot version `3.2.1`, including required dependency
upgrades to Spring Cloud `2023.0.0`, Spring Cloud AWS `3.1.0` and Spring
Cloud Azure `5.8.0`.
Other dependecies should be updated to bring everything to latest.

Refactor the SQS configuration, Spring Cloud AWS provides better
auto-configuration so the custom config class can be removed. The
message converter should be moved to ApplcationConfiguration.

Refactor FormBroadcastService to use the new builders for constructing
the PublishRequest and replace the SNS client and exceptions with their
new counterparts.

Refactor the SQS listener to removed the `deletionPolicy`, this is now
the default behaviour in Spring Cloud 3.x.x.

Refactor the FormService to use the new S3 client, several changes were
needed to account for the move away from an `S3Object` to instead return
wrapped responses with InputStreams or byte arrays for the content.
Remove the Commons IO dependency and init an auto-closed input stream
manually. This has the dual benefit of ensuring proper closer and
removing a dependency that is otherwise not used.
Rewrite the associated tests to use the new response object.

Remove the SentryConfiguration tests as the assertions were never
previously being ran due to a NoOp scope being used when the
configuration adds the context. 7.0.0 introduced a change where
Sentry.withScope will always run the callback, resulting in the test
now failing when running the assertions that never previously ran.

TIS21-SHED